miner, txpool: add a reserved gas constant for system transactions (#364)
* Revert "Add reserved gas for system transactions (#187)"

This partially reverts commit a1f364576eb0fd5f2dbb52c5fe23acb3fb2c9876 to make
the reserved gas a constant not a provided configuration from user. This still
keeps the created flag in that commit for compatibility.

* miner, txpool: add a reserved gas constant for system transactions

This commit adds the reserved gas for system transaction, which is previously a
user provided flag, as a constant.

In the miner part, the miner stops commit normal transactions to block when
block gas used is over the block's gas limit - reserved gas. So when finalizing
and including system transactions into block, the block's gas used does not exceed
the block's gas limit.

In the txpool part, when validating a transaction before putting it into the
pool, previously, we only check if the transaction's gas limit is not over the
block's gas limit. However, if the transaction's gas limit is over the block's
gas limit - reserved gas, it is never included into the block but still always
stay in pending transaction pool. This commit fixes the check to reject the
transaction with gas limit over the block's gas limit - reserved gas.